Rough Conversion **Needs work**
I used pandoc to do a conversion from the Thesis Template.docx file to an RMD and this is what was output.  When you knit it back into a word file the formatting is all sorts of messed up.  It needs a lot of cleaning up, but if we can get it to be exactly like the original template it would serve as the template for anyone wanting to write their thesis in markdown.